+++ flutter_inappwebview_forge_ios/lib/src/container_controller.dart
@@ -77,4 +77,11 @@ class IOSContainerController extends PlatformContainerController {
         'containerId': containerId,
       }) ??
       false;
+
+  @override
+  Future<bool> clearContainerData(String containerId) async =>
+      await _channel.invokeMethod<bool>('clearContainerData', <String, dynamic>{
+        'containerId': containerId,
+      }) ??
+      false;
 }